+++ Assets/com.gdcbd.bossbattle/Scenes/PlayerTestingScene.unity
@@ -1631,7 +1631,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 746588984}
-  - component: {fileID: 746588982}
+  - component: {fileID: 746588985}
   m_Layer: 0
   m_Name: ProjectileSpawner
   m_TagString: Untagged
@@ -1639,22 +1639,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &746588982
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 746588981}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6c2e5d109e85c4e24b0cfa81397bc4e2, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ProjectilePoolType: 0
-  CollectionChecks: 1
-  MaxPoolSize: 20
-  _bulletPrefab: {fileID: 3335241121232941687, guid: 74a0726c2900d4be699cc9db27a9f2d4, type: 3}
 --- !u!4 &746588984
 Transform:
   m_ObjectHideFlags: 0
@@ -1670,6 +1654,22 @@ Transform:
   m_Children: []
   m_Father: {fileID: 711255630}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &746588985
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 746588981}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: daaa6a0dfcd8a45b2b24eeb916619789, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObjectTypePool: 0
+  CollectionChecks: 1
+  MaxPoolSize: 20
+  _bulletPrefab: {fileID: 3335241121232941687, guid: 74a0726c2900d4be699cc9db27a9f2d4, type: 3}
 --- !u!1001 &770836608
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -5022,7 +5022,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7252581455397248090, guid: a9d4a71d420814da4b69a9a26903e113, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
