--- Assets/com.gdcbd.bossbattle/Scripts/Components/NEW OBJECT POOL/ProjectileSpawner.cs
@@ -1,12 +1,0 @@
-using UnityEngine;
-
-public class ProjectileSpawner : ObjectSpawner<ProjectileSpawner>
-{
-    [SerializeField] private GameObject _bulletPrefab;
-
-    protected override GameObject Prefab => _bulletPrefab;
-
-    protected override void Initialize()
-    {
-    }
-}
