+++ Assets/com.gdcbd.bossbattle/Scripts/Components/PoolObject.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 214068787a6144bb999897ea3748c0c4
+guid: 1c6f5113070d642af8d48ee23586d478
