+++ Assets/com.gdcbd.bossbattle/Scripts/Components/ProjectileSpawner.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace com.gdcbd.bossbattle.components
+{
+    public class ProjectileSpawner : ObjectSpawner<ProjectileSpawner>
+    {
+        [SerializeField] private GameObject _bulletPrefab;
+
+        protected override GameObject Prefab => _bulletPrefab;
+
+        protected override void Initialize()
+        {
+        }
+    }
+}
