+++ Assets/com.gdcbd.bossbattle/Scripts/Components/ProjectileSpawner.cs.meta
